_backend/requirements.txt
fastapi[all]
passlib[bcrypt]
PyJWT
pydantic-settings
python-dotenv
pymongo